requirements.txt
anthropic>=0.40.0
pyyaml>=6.0
textstat>=0.7.3
python-slugify>=8.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
